requirements.txt
aiofiles==0.8.0 ; python_version >= "3.10" and python_version < "4.0"
aiosqlite==0.17.0 ; python_version >= "3.10" and python_version < "4.0"
alembic==1.9.0 ; python_version >= "3.10" and python_full_version < "4.0.0"
anyio==3.6.2 ; python_version >= "3.10" and python_version < "4.0"
apscheduler==3.9.1.post1 ; python_version >= "3.10" and python_version < "4"
beautifulsoup4==4.11.1 ; python_version >= "3.10" and python_full_version < "4.0.0"
bidict==0.22.0 ; python_version >= "3.10" and python_full_version < "4.0.0"
bs4==0.0.1 ; python_version >= "3.10" and python_full_version < "4.0.0"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "4"
charset-normalizer==2.1.1 ; python_version >= "3.10" and python_version < "4"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32" or python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
expiringdict==1.2.2 ; python_version >= "3.10" and python_full_version < "4.0.0"
fastapi==0.88.0 ; python_version >= "3.10" and python_version < "4.0"
feedparser==6.0.10 ; python_version >= "3.10" and python_full_version < "4.0.0"
greenlet==2.0.1 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
h2==4.1.0 ; python_version >= "3.10" and python_version < "4.0"
hpack==4.0.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==0.16.2 ; python_version >= "3.10" and python_version < "4.0"
httptools==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.23.1 ; python_version >= "3.10" and python_full_version < "4.0.0"
httpx[http2]==0.23.1 ; python_version >= "3.10" and python_version < "4.0"
hyperframe==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
loguru==0.6.0 ; python_version >= "3.10" and python_version < "4.0"
mako==1.2.4 ; python_version >= "3.10" and python_full_version < "4.0.0"
markdown==3.4.1 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.1 ; python_version >= "3.10" and python_version < "4.0"
msgpack==1.0.4 ; python_version >= "3.10" and python_version < "4.0"
multidict==6.0.3 ; python_version >= "3.10" and python_version < "4.0"
nonebot-adapter-onebot==2.1.5 ; python_version >= "3.10" and python_version < "4.0"
nonebot-bison==0.6.3 ; python_version >= "3.10" and python_full_version < "4.0.0"
nonebot-plugin-apscheduler==0.2.0 ; python_version >= "3.10" and python_version < "4.0"
nonebot-plugin-atri==0.0.4 ; python_version >= "3.10" and python_version < "4.0"
nonebot-plugin-chatgpt==0.7.2 ; python_version >= "3.10" and python_full_version < "4.0.0"
nonebot-plugin-datastore==0.4.0 ; python_version >= "3.10" and python_version < "4.0"
nonebot-plugin-htmlrender==0.2.0.1 ; python_version >= "3.10" and python_version < "4.0"
nonebot-plugin-localstore==0.2.0 ; python_version >= "3.10" and python_version < "4.0"
nonebot2==2.0.0rc2 ; python_version >= "3.10" and python_version < "4.0"
nonebot2[httpx]==2.0.0rc2 ; python_version >= "3.10" and python_version < "4.0"
openaiauth==0.0.6 ; python_version >= "3.10" and python_full_version < "4.0.0"
pillow==9.3.0 ; python_version >= "3.10" and python_full_version < "4.0.0"
playwright==1.28.0 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.2 ; python_version >= "3.10" and python_version < "4.0"
pydantic[dotenv]==1.10.2 ; python_version >= "3.10" and python_version < "4.0"
pyee==9.0.4 ; python_version >= "3.10" and python_version < "4.0"
pygments==2.13.0 ; python_version >= "3.10" and python_version < "4.0"
pygtrie==2.5.0 ; python_version >= "3.10" and python_version < "4.0"
pyjwt==2.6.0 ; python_version >= "3.10" and python_full_version < "4.0.0"
pymdown-extensions==9.9 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
python-engineio==4.3.4 ; python_version >= "3.10" and python_full_version < "4.0.0"
python-markdown-math==0.8 ; python_version >= "3.10" and python_version < "4.0"
python-socketio==5.7.2 ; python_version >= "3.10" and python_full_version < "4.0.0"
pytz-deprecation-shim==0.1.0.post0 ; python_version >= "3.10" and python_version < "4"
pytz==2022.6 ; python_version >= "3.10" and python_version < "4"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.28.1 ; python_version >= "3.10" and python_version < "4"
rfc3986[idna2008]==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
setuptools==65.6.3 ; python_version >= "3.10" and python_version < "4"
sgmllib3k==1.0.0 ; python_version >= "3.10" and python_full_version < "4.0.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
soupsieve==2.3.2.post1 ; python_version >= "3.10" and python_full_version < "4.0.0"
sqlalchemy2-stubs==0.0.2a29 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==1.4.41 ; python_version >= "3.10" and python_version < "4.0"
sqlmodel==0.0.8 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.22.0 ; python_version >= "3.10" and python_version < "4.0"
tinydb==4.7.0 ; python_version >= "3.10" and python_version < "4.0"
tls-client==0.1.5 ; python_version >= "3.10" and python_full_version < "4.0.0"
tomlkit==0.11.6 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.4.0 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2022.7 ; python_version >= "3.10" and python_version < "4"
tzlocal==4.2 ; python_version >= "3.10" and python_version < "4"
urllib3==1.26.13 ; python_version >= "3.10" and python_version < "4"
uvicorn[standard]==0.20.0 ; python_version >= "3.10" and python_version < "4.0"
uvloop==0.17.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "4.0"
watchfiles==0.18.1 ; python_version >= "3.10" and python_version < "4.0"
websockets==10.4 ; python_version >= "3.10" and python_version < "4.0"
win32-setctime==1.1.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
yarl==1.8.2 ; python_version >= "3.10" and python_version < "4.0"
